Finalize character files

+++ DB.SqlServer/Tables/CharacterClassification.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[CharacterClassification]
 (
-	[CharacterClassificationId] INT NOT NULL IDENTITY(1,1),
+	[CharacterClassificationId] TINYINT NOT NULL IDENTITY(1,1),
 	[Classification] VARCHAR(50) NOT NULL,
     CONSTRAINT [PK_CharacterClassification] PRIMARY KEY ([CharacterClassificationId])
 )
